refactor(supervisor): move tray _run behind the per-OS backend seam
